@@ -79,16 +79,13 @@ def test_extract_migration():
         // add your reverse migration here
     }'''
 
-    script = '''
+    script = f'''
     test string please ignore
     {forward_function}
     also this
     {reverse_function}
     more garbage
-    '''.format(
-        forward_function=forward_function,
-        reverse_function=reverse_function
-    )
+    '''
 
     forward_migration = extract_migration(script, 'forward')
     reverse_migration = extract_migration(script, 'reverse')
